+++ client/src/util/gameUtils/checkScreenScale.js
