+++ client/src/util/gameUtils/getDistance.js
