+++ client/src/util/updateObjects/updateAsteroids.js
